changing from float to int for memory
